Move entity mapping information out of webmvc package
Mapping information is not tied directly to webmvc, but is more generic functionality
