+++ contentgrid-spring-data-rest/src/main/java/com/contentgrid/spring/data/rest/mapping/persistent/PersistentPropertyProperty.java
@@ -1,8 +1,8 @@
-package com.contentgrid.spring.data.rest.webmvc.mapping.persistent;
+package com.contentgrid.spring.data.rest.mapping.persistent;
 
-import com.contentgrid.spring.data.rest.webmvc.mapping.Container;
-import com.contentgrid.spring.data.rest.webmvc.mapping.Property;
-import com.contentgrid.spring.data.rest.webmvc.mapping.typeinfo.TypeInformationContainer;
+import com.contentgrid.spring.data.rest.mapping.Container;
+import com.contentgrid.spring.data.rest.mapping.Property;
+import com.contentgrid.spring.data.rest.mapping.typeinfo.TypeInformationContainer;
 import java.lang.annotation.Annotation;
 import java.util.Optional;
 import java.util.function.Predicate;
