+++ contentgrid-spring-data-rest/src/main/java/com/contentgrid/spring/data/rest/mapping/typeinfo/TypeInformationProperty.java
@@ -1,7 +1,7 @@
-package com.contentgrid.spring.data.rest.webmvc.mapping.typeinfo;
+package com.contentgrid.spring.data.rest.mapping.typeinfo;
 
-import com.contentgrid.spring.data.rest.webmvc.mapping.Container;
-import com.contentgrid.spring.data.rest.webmvc.mapping.Property;
+import com.contentgrid.spring.data.rest.mapping.Container;
+import com.contentgrid.spring.data.rest.mapping.Property;
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Field;
 import java.util.Optional;
